Fixed the corrections
