+++ lib/src/features/onboarding/views/onboarding_v.dart
@@ -56,7 +56,7 @@ class OnBoardingView extends ConsumerWidget {
                     isEnabled: true,
                     buttonTitle: 'Open an account',
                     onPressed: () =>
-                        FNavigator.navigateToRoute(const OpenAccount())),
+                        FNavigator.navigateToRoute( OpenAccount())),
               ),
               addVerticalSpacing(10),
               Padding(
